Refine the ManifestBitmap API for manifest deletion vectors
Backs `RoaringManifestBitmap` with an `Either[Array[Byte], Seq[Int]]` so a bitmap
can be built either from raw positions or from already-serialized bytes, deferring
the decode into a `RoaringBitmapArray` until the first bit-level access. The two
factory methods are `fromPositions` and `fromSerializedByteArray`.

`serializeAsByteArray` returns the original bytes when the bitmap was never mutated
(with a defensive copy so callers cannot mutate the internal buffer), and re-encodes
from the decoded bitmap otherwise. `BackReference.pos` narrows from `Long` to `Int`,
and the manifest write path (`AMTWriteHelper`, `IncrementalAMTWriter`) plus the
associated tests are updated to the new API.

+++ spark/src/test/scala/org/apache/spark/sql/delta/amt/AMTBackReferenceSuite.scala
@@ -144,7 +144,7 @@ class AMTBackReferenceSuite extends AMTCheckpointTestBase with DeletionVectorsTe
       assert(leafPaths.contains(br.manifest),
         s"Back-ref manifest ${br.manifest} must be one of the tree's leaves $leafPaths.")
       // The back-ref must point at exactly the leaf entry describing this data file.
-      assert(groundTruth.get((br.manifest, br.pos)).contains(add.path),
+      assert(groundTruth.get((br.manifest, br.pos.toLong)).contains(add.path),
         s"Back-ref (${br.manifest}, ${br.pos}) must resolve to the entry for ${add.path}.")
     }
   }
@@ -433,7 +433,7 @@ class AMTBackReferenceSuite extends AMTCheckpointTestBase with DeletionVectorsTe
   test("commit fails when a present file's tombstone carries the wrong back reference") {
     withTable("amt_commit_wrong") {
       val adds = emitStampedAddFiles("amt_commit_wrong")
-      val wrongBr = adds.head.backReference.get.copy(pos = adds.head.backReference.get.pos + 1000L)
+      val wrongBr = adds.head.backReference.get.copy(pos = adds.head.backReference.get.pos + 1000)
       val wrong = adds.head.removeWithTimestamp().copy(backReference = Some(wrongBr))
       val ex = intercept[IllegalStateException] {
         commitActions("amt_commit_wrong", Seq(wrong))
